+++ lib/src/map_widget.dart
@@ -44,6 +44,8 @@ class MapWidget extends StatefulWidget {
   const MapWidget({
     super.key,
     this.mapOptions,
+    @Deprecated(
+        'Use [viewport] to specify the camera position and behavior of the map')
     this.cameraOptions,
     // FIXME Flutter 3.x has memory leak on Android using in SurfaceView mode, see https://github.com/flutter/flutter/issues/118384
     // As a workaround default is true.
@@ -66,7 +68,9 @@ class MapWidget extends StatefulWidget {
     this.onStyleImageMissingListener,
     this.onStyleImageUnusedListener,
     this.onResourceRequestListener,
+    @Deprecated('Use [MapboxMap.addInteraction] instead')
     this.onTapListener,
+    @Deprecated('Use [MapboxMap.addInteraction] instead')
     this.onLongTapListener,
     this.onScrollListener,
     this.onZoomListener,
@@ -77,6 +81,8 @@ class MapWidget extends StatefulWidget {
   final MapOptions? mapOptions;
 
   /// The Initial Camera options when creating a MapWidget.
+  @Deprecated(
+      'This will be removed in future major version, use [viewport] instead')
   final CameraOptions? cameraOptions;
 
   /// Flag indicating to use a TextureView as render surface for the MapWidget.
@@ -175,7 +181,9 @@ class MapWidget extends StatefulWidget {
   /// ```
   final ViewportState? viewport;
 
+  @Deprecated('Use [MapboxMap.addInteraction] instead')
   final OnMapTapListener? onTapListener;
+  @Deprecated('Use [MapboxMap.addInteraction] instead')
   final OnMapLongTapListener? onLongTapListener;
   final OnMapScrollListener? onScrollListener;
   final OnMapZoomListener? onZoomListener;